Medical/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
weaviate-client==3.25.2
huggingface_hub==0.16.4
sentence-transformers==2.2.2
numpy==1.26.0